backend/requirements/base.txt
Django==5.2.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.11
stripe==15.0.0
Pillow==12.0.0
python-decouple==3.8
dj-database-url==3.1.2
gunicorn==23.0.0
